Data Wrangling Project
This project assess red and white wine and the physicochemical properties that make up wine and the correlation between those properties and quality scores. It was observed that alcohol content is a has a positive correlation to wine quality in both red and white wines. It was also observed that volatile acidity had a negative correlation to quality especially in red wine types. Included in the file is the project notebook, links for the data, the raw data, the combined cleaned data, and the combined cleaned long format data. The visualizations used were boxplots and heatmaps. The purpose of the project was to demonstrate data wrangling capabilities, being able to identify tidiness and quality issues with data and cleaning them to perform analysis and develop insights.